adding testing controller and config
