+++ swabbie-core/src/main/kotlin/com/netflix/spinnaker/config/TestingConfiguration.kt
@@ -1,0 +1,35 @@
+/*
+ *
+ *  * Copyright 2018 Netflix, Inc.
+ *  *
+ *  * Licensed under the Apache License, Version 2.0 (the "License")
+ *  * you may not use this file except in compliance with the License.
+ *  * You may obtain a copy of the License at
+ *  *
+ *  *   http://www.apache.org/licenses/LICENSE-2.0
+ *  *
+ *  * Unless required by applicable law or agreed to in writing, software
+ *  * distributed under the License is distributed on an "AS IS" BASIS,
+ *  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  * See the License for the specific language governing permissions and
+ *  * limitations under the License.
+ *
+ */
+
+package com.netflix.spinnaker.config
+
+import com.netflix.spinnaker.swabbie.model.AlwaysCleanRule
+import org.springframework.boot.autoconfigure.condition.ConditionalOnExpression
+import org.springframework.context.annotation.Bean
+import org.springframework.context.annotation.Configuration
+
+@Configuration
+@ConditionalOnExpression("\${swabbie.testing.enabled:false}")
+open class TestingConfiguration {
+
+  @ConditionalOnExpression("\${swabbie.testing.alwaysCleanRuleConfig.enabled:false}")
+  @Bean
+  open fun alwaysCleanRule(swabbieProperties: SwabbieProperties) =
+    AlwaysCleanRule(swabbieProperties)
+
+}
